requirements.txt
python-dotenv
python-telegram-bot[job-queue]
playwright
requests
gologin